Sample validation
Write a function that examines the barcode pairs of each sample in a
given json of the above structure and returns false if any of the
barcode pairs in one sample are the same as any of the barcode pairs in
another sample. Barcode pairs are considered the same if both parts of
the barcodes are the same or if one part or both parts of the barcode
pair differ by only one base (character) from the same part of the
barcode pair in another sample. (max 30 lines)